RDKCOM-1949 RDKBCMF-255 CcspTr069Pa build errors
Emulator build is failing when all warnings are treated as errors in
meta-rdk-broadband/52095.
REFPLTB-984: EMULATOR CcspTr069Pa segfaulting with fix for treating warnings as errors
Reason for change: Emulator build is failing when all warnings are treated as errors in
meta-rdk-broadband/52095
Test procedure: systemctl status CcspTr069
Risks: Low
Test Procedure: None.
(cherry picked from commit e88aa431f6f824744b99cebe940b3929bc3cbb42)

+++ source-pc/include/ccsp_cwmp_helper_api.h
@@ -1019,7 +1019,7 @@ CcspMemorySearch
             return  NULL;
         }
 
-        if ( pBufEnd - pNext + 1 >= ulPatternLen &&
+        if ( (ULONG)(pBufEnd - pNext + 1) >= ulPatternLen &&
              AnscEqualString2((char*)pNext, (char*)pPattern, ulPatternLen, bCaseSensitive) )
         {
             return  (char*)pNext;
